Integrated Components
Integrated newly created components into the app.
