+++ src/Components/AppNavigator.js
@@ -1,71 +1,81 @@
 import React from 'react';
+
+import { Button, Text, Alert } from 'react-native';
+
 import { NavigationContainer } from '@react-navigation/native';
 import { createStackNavigator } from '@react-navigation/stack';
 
 import DashboardScreen from '../Screens/HMI/DashboardScreen';
 import RecordingScreen from '../Screens/HMI/RecordingScreen';
 import TripResultScreen from '../Screens/HMI/TripResultScreen';
-import CognitiveTestScreen from '../Screens/HMI/CognitiveTestScreen';
-import RegistrationForm from '../Screens/RegistrationForm';
-import LoginScreen from '../Screens/LoginScreen'; // Import the Login screen
+import CognitiveTestScreen from '../Screens/DriveTests/CognitiveTestScreen';
+import RegistrationForm from '../Screens/Authentication/RegistrationForm';
+import LoginScreen from '../Screens/Authentication/LoginScreen';
+
+import { colors } from '../Styles/StyleSheet';
 
 const Stack = createStackNavigator();
 
 const AppNavigator = () => {
   return (
     <NavigationContainer>
-      <Stack.Navigator initialRouteName="Registration"> {/* Set Registration as the initial route */}
-        {/* Registration Screen */}
+      <Stack.Navigator initialRouteName="Login">
         <Stack.Screen
           name="Registration"
           component={RegistrationForm}
           options={{
-            title: 'Register', // Title for the Registration screen
+            title: '',
+            headerLeft: () => null, // This hides the back button
           }}
         />
-        
-        {/* Login Screen */}
         <Stack.Screen
           name="Login"
           component={LoginScreen}
           options={{
-            title: 'Login', // Title for the Login screen
+            title: '',
+            headerLeft: () => null, // This hides the back button
           }}
         />
-        
-        {/* Dashboard Screen */}
         <Stack.Screen
           name="Dashboard"
           component={DashboardScreen}
-          options={{
-            title: 'Dashboard', // Title for the Dashboard screen
-          }}
+          options={({ navigation }) => ({
+            title: 'Dashboard',
+            headerLeft: () => null, // This hides the back button
+            headerRight: () => (
+              <Button
+                onPress={() => {
+                  Alert.alert('Successfully Logged Out.');
+                  navigation.navigate('Login');
+                }}
+                title="Logout"
+                color={colors.error}
+              />
+            ),
+          })}
         />
-        
-        {/* Cognitive Test Screen */}
         <Stack.Screen
           name="CognitiveTest"
           component={CognitiveTestScreen}
           options={{
-            title: 'Cognitive Test', // Title for the Cognitive Test screen
+            title: 'CognitiveTest', 
+            headerLeft: () => null, // This hides the back button
           }}
         />
-        
-        {/* Recording Screen */}
         <Stack.Screen
           name="Recording"
           component={RecordingScreen}
           options={{
-            title: 'Recording', // Title for the Recording screen
+            title: 'Recording', 
+            headerLeft: () => null, // This hides the back button
           }}
         />
-        
-        {/* Trip Results Screen */}
         <Stack.Screen
           name="TripResultScreen"
           component={TripResultScreen}
           options={{
-            title: 'Trip Results', // Title for the Trip Results screen
+            title: 'Trip Results', 
+            headerLeft: () => null, // This hides the back button
           }}
         />
       </Stack.Navigator>
